frontend/admin_ui.py
import streamlit as st
from google.cloud import storage
import os

st.title("Customer Support - Admin Portal")

# GCS setup
os.environ["GOOGLE_APPLICATION_CREDENTIALS"] = "path/to/credentials.json"  # Placeholder
client = storage.Client()
bucket = client.bucket("your-bucket-name")

st.header("Upload Files")

uploaded_file = st.file_uploader("Choose a file", type=["pdf", "xlsx", "eml"])
if uploaded_file is not None:
    # Upload to GCS
    blob = bucket.blob(uploaded_file.name)
    blob.upload_from_file(uploaded_file)
    st.success("File uploaded to GCS. Processing will be triggered automatically.")

# Chat interface (same as user)
st.header("Chat with Support")
# Similar to user_ui, but perhaps admin can view all or something. For now, same.