frontend/requirements.txt
streamlit==1.28.1
requests==2.31.0
google-cloud-storage==2.10.0